Add files via upload
This system, a RFID reader is used at the entry point for user authentication. A GSM module is used so that the parking slot no, the duration of parking and the fee deducted from his card is sent to the user when he exits the parking space. The lights in the parking space glow only when the vehicle is present.